refactor: extract useAmountBlur hook
